Add fzf vim options
The fzf search seems to have a better ranking and I can now do a live
search with :Ag.

The denite window position has been changed to bottom to be consistent
with fzf and to stay close to the prompt. (Yes I know you can change the
position of the fzf window but you cannot change the denite prompt.)