Switch from ycm to deoplete
The deoplete solutions are a little bit more plug-and-play as they do
not require compilation and offer more options for customisations.

C++ completion is offered by deoplete-clang. deoplete-clang2 gives worse
results for some reason.
Javascript completion is done with ternjs.
echodoc shows the function signature while typing in the arguments, but
requires the removal of the mode indicator.

With neopairs and converter_auto_paren the function parens are added
automatically.